[ALPS04170455] RTC: Fix code warning
Use unsigned variable for & and | operator.
No Infinite loop.
Check whether unsigned value is greater than zero is unnecessary.

MTK-Commit-Id: e1ce583b004e5c305ee7b3f1f7cad4f8a35bd774

CR-Id: ALPS04170455
Feature: Power-Off Alarm
(cherry picked from commit 3bb7b3be65a4290d6076f2fb63dc177e3ec86d1b)
(cherry-pick from a2ff643d1469b14da8801422ef6ac1cfa4cd1314)

+++ drivers/misc/mediatek/rtc/mtk_rtc_hal_common.c
@@ -116,7 +116,7 @@ void hal_rtc_set_spare_register(enum rtc_spare_enum cmd, u16 val)
 {
 	u16 tmp_val;
 
-	if (cmd >= 0 && cmd < RTC_SPAR_NUM) {
+	if ((int)cmd >= 0 && cmd < RTC_SPAR_NUM) {
 		hal_rtc_xinfo("%s: cmd[%d], set rg[0x%x, 0x%x , %d] = 0x%x\n",
 				__func__, cmd,
 				rtc_spare_reg[cmd][RTC_REG],
@@ -139,7 +139,7 @@ u16 hal_rtc_get_spare_register(enum rtc_spare_enum cmd)
 {
 	u16 tmp_val;
 
-	if (cmd >= 0 && cmd < RTC_SPAR_NUM) {
+	if ((int)cmd >= 0 && cmd < RTC_SPAR_NUM) {
 
 		tmp_val = rtc_read(rtc_spare_reg[cmd][RTC_REG]);
 		tmp_val =
@@ -208,29 +208,31 @@ void hal_rtc_get_alarm_time(struct rtc_time *tm)
 void hal_rtc_set_alarm_time(struct rtc_time *tm)
 {
 	hal_rtc_xinfo("mon = %d, day = %d, hour = %d\n",
-		      (rtc_read(RTC_AL_MTH) & ~(RTC_AL_MTH_MASK)) | tm->tm_mon,
-		      (rtc_read(RTC_AL_DOM) & ~(RTC_AL_DOM_MASK)) | tm->tm_mday,
-		      (rtc_read(RTC_AL_HOU) & ~(RTC_AL_HOU_MASK)) | tm->
-		      tm_hour);
+		      (rtc_read(RTC_AL_MTH) & ~(RTC_AL_MTH_MASK)) |
+		      (unsigned int)tm->tm_mon,
+		      (rtc_read(RTC_AL_DOM) & ~(RTC_AL_DOM_MASK)) |
+		      (unsigned int)tm->tm_mday,
+		      (rtc_read(RTC_AL_HOU) & ~(RTC_AL_HOU_MASK)) |
+		      (unsigned int)tm->tm_hour);
 
 	rtc_write(RTC_AL_YEA,
 		  (rtc_read(RTC_AL_YEA) & ~(RTC_AL_YEA_MASK)) |
-		  (tm->tm_year & RTC_AL_YEA_MASK));
+		  ((unsigned int)tm->tm_year & RTC_AL_YEA_MASK));
 	rtc_write(RTC_AL_MTH,
 		  (rtc_read(RTC_AL_MTH) & ~(RTC_AL_MTH_MASK)) |
-		  (tm->tm_mon & RTC_AL_MTH_MASK));
+		  ((unsigned int)tm->tm_mon & RTC_AL_MTH_MASK));
 	rtc_write(RTC_AL_DOM,
 		  (rtc_read(RTC_AL_DOM) & ~(RTC_AL_DOM_MASK)) |
-		  (tm->tm_mday & RTC_AL_DOM_MASK));
+		  ((unsigned int)tm->tm_mday & RTC_AL_DOM_MASK));
 	rtc_write(RTC_AL_HOU,
 		  (rtc_read(RTC_AL_HOU) & ~(RTC_AL_HOU_MASK)) |
-		  (tm->tm_hour & RTC_AL_HOU_MASK));
+		  ((unsigned int)tm->tm_hour & RTC_AL_HOU_MASK));
 	rtc_write(RTC_AL_MIN,
 		  (rtc_read(RTC_AL_MIN) & ~(RTC_AL_MIN_MASK)) |
-		  (tm->tm_min & RTC_AL_MIN_MASK));
+		  ((unsigned int)tm->tm_min & RTC_AL_MIN_MASK));
 	rtc_write(RTC_AL_SEC,
 		  (rtc_read(RTC_AL_SEC) & ~(RTC_AL_SEC_MASK)) |
-		  (tm->tm_sec & RTC_AL_SEC_MASK));
+		  ((unsigned int)tm->tm_sec & RTC_AL_SEC_MASK));
 	rtc_write(RTC_AL_MASK, RTC_AL_MASK_DOW);	/* mask DOW */
 	rtc_write_trigger();
 }
@@ -268,30 +270,33 @@ void hal_rtc_set_pwron_alarm_time(struct rtc_time *tm)
 {
 	rtc_write(RTC_PWRON_YEA,
 		  (rtc_read(RTC_PWRON_YEA) & ~(RTC_PWRON_YEA_MASK))
-		  | ((tm->tm_year << RTC_PWRON_YEA_SHIFT) &
+		  | (((unsigned int)tm->tm_year << RTC_PWRON_YEA_SHIFT) &
 		     RTC_PWRON_YEA_MASK));
 	rtc_write_trigger();
 	rtc_write(RTC_PWRON_MTH,
 		  (rtc_read(RTC_PWRON_MTH) & ~(RTC_PWRON_MTH_MASK))
-		  | ((tm->tm_mon << RTC_PWRON_MTH_SHIFT) & RTC_PWRON_MTH_MASK));
+		  | (((unsigned int)tm->tm_mon << RTC_PWRON_MTH_SHIFT) &
+		     RTC_PWRON_MTH_MASK));
 	rtc_write_trigger();
 	rtc_write(RTC_PWRON_DOM,
 		  (rtc_read(RTC_PWRON_DOM) & ~(RTC_PWRON_DOM_MASK))
-		  | ((tm->tm_mday << RTC_PWRON_DOM_SHIFT) &
+		  | (((unsigned int)tm->tm_mday << RTC_PWRON_DOM_SHIFT) &
 		     RTC_PWRON_DOM_MASK));
 	rtc_write_trigger();
 	rtc_write(RTC_PWRON_HOU,
 		  (rtc_read(RTC_PWRON_HOU) & ~(RTC_PWRON_HOU_MASK))
-		  | ((tm->tm_hour << RTC_PWRON_HOU_SHIFT) &
+		  | (((unsigned int)tm->tm_hour << RTC_PWRON_HOU_SHIFT) &
 		     RTC_PWRON_HOU_MASK));
 	rtc_write_trigger();
 	rtc_write(RTC_PWRON_MIN,
 		  (rtc_read(RTC_PWRON_MIN) & ~(RTC_PWRON_MIN_MASK))
-		  | ((tm->tm_min << RTC_PWRON_MIN_SHIFT) & RTC_PWRON_MIN_MASK));
+		  | (((unsigned int)tm->tm_min << RTC_PWRON_MIN_SHIFT) &
+		     RTC_PWRON_MIN_MASK));
 	rtc_write_trigger();
 	rtc_write(RTC_PWRON_SEC,
 		  (rtc_read(RTC_PWRON_SEC) & ~(RTC_PWRON_SEC_MASK))
-		  | ((tm->tm_sec << RTC_PWRON_SEC_SHIFT) & RTC_PWRON_SEC_MASK));
+		  | (((unsigned int)tm->tm_sec << RTC_PWRON_SEC_SHIFT) &
+		     RTC_PWRON_SEC_MASK));
 	rtc_write_trigger();
 }
